Source/requirements.txt
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0
PyQt6-QScintilla>=2.14.1
jedi>=0.19.0
pygments>=2.16.0
python-lsp-server[all]>=1.6.1
autopep8>=2.0.0
pyflakes>=3.1.0
pycodestyle>=2.11.0
GitPython>=3.1.40
